refactor to use tcprouterclient
